Move gzdoom_main_wrapper.hpp from gvizdoom_client to gvizdoom
